:recycle: Rename to setting/setting_page_state
