+++ lib/domain/settings/setting_page_state.dart
@@ -1,7 +1,7 @@
 import 'package:pilll/entity/setting.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'setting.freezed.dart';
+part 'setting_page_state.freezed.dart';
 
 @freezed
 abstract class SettingState implements _$SettingState {
